Utils#playlist: filter out null songs correctly
Previously, the code was `.filter`ing on the array of *promises*, so
because all objects are truthy the filter wasn't actually doing
anything. This moves the `filter` call to after the `await
Promise.all(...)` so it properly filters the `Song | null` returned from
the promises.

This commit also removes some unnecessary type assertions.